requirements.txt
python-dotenv
streamlit
google-generativeai
ipython
pillow

python-dotenv==1.0.1
streamlit==1.38.0
google-generativeai==0.5.2
ipython==8.27.0
pillow==10.3.0
altair==5.4.0
numpy==2.1.0
pandas==2.2.2
...
